+++ SD_StudentLoanServices/src/main/java/com/sd/sls/loanapplication/status/state/WithdrawState.java
@@ -9,13 +9,13 @@ import com.sd.sls.loanapplication.dao.ILoanApplicationDAO;
  * @Author: Abhishek Vishwakarma
  */
 @Component
-public class WithdrawState implements IApplicationStatusState {
-
+public class WithdrawState implements IApplicationStatusState 
+{
 	@Autowired
 	private ILoanApplicationDAO loanApplicationDAO;
 
 	@Override
-	public int updateStatus(Long loanApplicationId) {
+	public int updateStatus(int loanApplicationId) {
 		return loanApplicationDAO.withdrawApplication(loanApplicationId);
 	}
 }
